Fixed Broken Code
I downloaded the code and loaded it in the browser. The first error was in the index.cfm file on line 24. There was a missing closing </cfif> tag. The next error was item is  not valid in cfloop with an array. I changed it to index. I added a semi-colon to the end of line 2 in the application.cfm file.  The JSON file was missing a comma so I added that. I experienced an issue locally because I put the code in a folder that was not the root, so to further troubleshoot I had to change the path for the cfhttp call. I changed it back before submitting.  The next issue was that a Struct type cannot be used as an array. I updated the reference to be an array by changing it to Animals.Animals. Next the form variable qCheck was undefined. I uncommented that line of code. Finally the match variable was undefined outside of a cfiif block. I defined it outside the block. I used the stack trace to debug as well as dumping and aborting to see what the variables actually contained. I completed the fixes in about 35-40 minutes. I did experience some issues pushing the code back to GitHub due to the way my SourceTree client is setup for my current company. I downloaded another client and started from scratch.